+++ bar_chart_race/_line_chart_race.py
@@ -24,7 +24,7 @@ class _LineChartRace(CommonChart):
                  end_period_pause, period_summary_func, line_width_data, agg_line_func, 
                  agg_line_kwargs, others_line_func, others_line_kwargs, fade, min_fade, 
                  images, colors, title, line_label_font, tick_label_font, tick_template, 
-                 shared_fontdict, scale, fig, writer, line_kwargs, fig_kwargs, filter_column_colors):
+                 shared_fontdict, scale, fig, writer, line_kwargs, fig_kwargs):
         self.filename = filename
         self.extension = self.get_extension()
         self.n_lines = n_lines or df.shape[1]
@@ -33,9 +33,9 @@ class _LineChartRace(CommonChart):
         self.end_period_pause = end_period_pause
         self.period_summary_func = period_summary_func
         self.agg_line_func = agg_line_func
-        self.agg_line_kwargs, self.agg_line_label = self.get_extra_line_kwargs(agg_line_kwargs, 'agg')
+        self.agg_line_kwargs, self.agg_line_label = self.get_line_kwargs(agg_line_kwargs, 'agg')
         self.others_line_func = others_line_func
-        self.others_line_kwargs, self.others_line_label = self.get_extra_line_kwargs(others_line_kwargs, 'others')
+        self.others_line_kwargs, self.others_line_label = self.get_line_kwargs(others_line_kwargs, 'others')
         self.line_width_data = self.get_line_width_data(line_width_data)
         self.fade = fade
         self.min_fade = min_fade
@@ -47,12 +47,11 @@ class _LineChartRace(CommonChart):
         self.scale = scale
         self.writer = self.get_writer(writer)
         self.fps = 1000 / self.period_length * steps_per_period
-        self.filter_column_colors = filter_column_colors
         self.validate_params()
 
-        self.line_kwargs = self.get_line_kwargs(line_kwargs)
+        self.line_kwargs = self.get_line_kwargs(line_kwargs, 'line')
         self.html = self.filename is None
-        self.df_values, self.df_ranks, self.df_others = self.prepare_data(df)
+        self.all_values, self.df_values, self.df_ranks, self.df_others, self.others_agg_line = self.prepare_data(df)
         self.agg_line = self.prepare_agg_line()
         self.is_x_date = self.df_values.index.dtype.kind == 'M'
         self.colors = self.get_colors(colors)
@@ -63,16 +62,19 @@ class _LineChartRace(CommonChart):
         self.collections = {}
         self.other_collections = {}
         self.texts = {}
-        self.aggregate_others = isinstance(self.others_line_func, str) or callable(self.others_line_func)
         self.images = self.get_images(images)
         self.image_radius = self.fig.get_figwidth() * self.fig.dpi * .02
 
-    def get_extra_line_kwargs(self, kwargs, kind):
+    def get_line_kwargs(self, kwargs, kind):
+        defaults = {'lw': 1.5, 'ls': '-'}
         text = None
-        if kind == 'agg':
-            defaults = {'color': OTHERS_COLOR, 'lw': 1.5, 'ls': '-'}
-        else:
-            defaults = {'color': AGG_COLOR, 'lw': 1.5, 'ls': '-'}
+        if kind == 'others':
+            defaults['color'] = OTHERS_COLOR
+        elif kind == 'agg':
+            defaults['color'] = AGG_COLOR
+        elif kind == 'line':
+            defaults['alpha'] = 1
+
         if kwargs is None:
             kwargs = defaults
         elif isinstance(kwargs, dict):
@@ -84,10 +86,12 @@ class _LineChartRace(CommonChart):
         
         if 'linewidth' in kwargs:
             kwargs['lw'] = kwargs.pop('linewidth')
-        if 'c' in kwargs:
-            kwargs['color'] = kwargs.pop('c')
         if 'linestyle' in kwargs:
             kwargs['ls'] = kwargs.pop('linestyle')
+        if kind == 'line':
+            return kwargs
+        if 'c' in kwargs:
+            kwargs['color'] = kwargs.pop('c')
         
         kwargs['color'] = mcolors.to_rgba(kwargs['color'])
         return kwargs, text
@@ -110,16 +114,10 @@ class _LineChartRace(CommonChart):
         elif self.filename is not None:
             raise TypeError('`filename` must be None or a string')
 
-    def get_line_kwargs(self, line_kwargs):
-        line_kwargs = line_kwargs or {}
-        if 'alpha' not in line_kwargs:
-            line_kwargs['alpha'] = .8
-        return line_kwargs
-
     def get_font(self, font, ticks=False):
-        default_font_dict = {'size': 7}
+        default_font_dict = {'size': 7, 'ha': 'left', 'va': 'center'}
         if ticks:
-            default_font_dict['ha'] = 'right'
+            default_font_dict['ha'] = 'center'
 
         if font is None:
             font = default_font_dict
@@ -141,20 +139,16 @@ class _LineChartRace(CommonChart):
 
         idx = values.iloc[-1].sort_values(ascending=False).index
         top_cols, other_cols = idx[:self.n_lines], idx[self.n_lines:]
+        all_values = values.copy()
         values, ranks, others = values[top_cols], ranks[top_cols], values[other_cols]
 
-        if self.others_line_func is None or len(other_cols) == 0:
-            others = None
-        others = self.prepare_others_df(others)
-        return values, ranks, others
+        if self.others_line_func in (None, True) or len(other_cols) == 0:
+            others_agg_line = None
+        else:
+            others_agg_line = self.prepare_others_agg_line(others)
+        return all_values, values, ranks, others, others_agg_line
 
-    def prepare_others_df(self, others):
-        # when others_line_func is None -> others is None
-        # when others_line_func is True -> others is a dataframe
-        # when others_line_func is str or func -> others is a series
-        if others is None or self.others_line_func is True:
-            return others
-
+    def prepare_others_agg_line(self, others):
         if isinstance(self.others_line_func, str):
             s_others = others.agg(self.others_line_func, axis=1)
             label = self.others_line_func
@@ -172,10 +166,10 @@ class _LineChartRace(CommonChart):
         if self.agg_line_func is None:
             return
         if isinstance(self.agg_line_func, str):
-            s_agg = self.df_values.agg(self.agg_line_func, axis=1)
+            s_agg = self.all_values.agg(self.agg_line_func, axis=1)
             label = self.agg_line_func
         elif callable(self.agg_line_func):
-            s_agg = self.df_values.agg(self.agg_line_func, axis=1)
+            s_agg = self.all_values.agg(self.agg_line_func, axis=1)
             label = self.agg_line_func.__name__
         else:
             raise TypeError('`agg_line_func` must be either a string or function')
@@ -220,8 +214,6 @@ class _LineChartRace(CommonChart):
 
     def prepare_axes(self, ax):
         ax.grid(True, color='white')
-        if self.tick_template:
-            ax.yaxis.set_major_formatter(self.tick_template)
         ax.tick_params(labelsize=self.tick_label_font['size'], length=0, pad=2)
         ax.minorticks_off()
         ax.set_axisbelow(True)
@@ -230,12 +222,13 @@ class _LineChartRace(CommonChart):
         
         min_val = self.df_values.min().min()
         max_val = self.df_values.max().max()
-        if isinstance(self.df_others, pd.Series):
-            min_val = min(min_val, self.df_others.min())
-            max_val = max(max_val, self.df_others.max())
-        elif isinstance(self.df_others, pd.DataFrame):
+
+        if self.others_line_func is True:
             min_val = min(min_val, self.df_others.min().min())
             max_val = max(max_val, self.df_others.max().max())
+        elif self.others_agg_line is not None:
+            min_val = min(min_val, self.others_agg_line.min())
+            max_val = max(max_val, self.others_agg_line.max())
             
         if self.agg_line is not None:
             min_val = min(min_val, self.agg_line.min())
@@ -255,11 +248,17 @@ class _LineChartRace(CommonChart):
 
         ymin, ymax = ax.get_ylim()
         delta = (ymax - ymin) * .05
+        if self.scale == 'log':
+            delta = 0
+            ymax = 2 * ymax
         ax.set_ylim(ymin - delta, ymax + delta)
 
         if self.is_x_date:
             ax.xaxis_date()
             ax.xaxis.set_major_formatter(mdates.DateFormatter('%-m/%-d'))
+
+        if self.tick_template:
+            ax.yaxis.set_major_formatter(self.tick_template)
 
     def get_subplots_adjust(self):
         import io
@@ -278,6 +277,18 @@ class _LineChartRace(CommonChart):
         bottom = ax.get_position().y0 - ymin + .01
         return left, bottom
 
+    def get_fig(self, fig):
+        if fig is not None and not isinstance(fig, plt.Figure):
+            raise TypeError('`fig` must be a matplotlib Figure instance')
+        if fig is not None:
+            if not fig.axes:
+                raise ValueError('The figure passed to `fig` must have an axes')
+            ax = fig.axes[0]
+            self.prepare_axes(ax)
+        else:
+            fig = self.create_figure()
+        return fig
+
     def create_figure(self):
         fig = plt.Figure(**self.fig_kwargs)
         ax = fig.add_subplot()
@@ -290,15 +301,8 @@ class _LineChartRace(CommonChart):
         if images is None:
             return
         if isinstance(images, str):
-            if images == 'country':
-                url = 'https://github.com/hjnilsson/country-flags/raw/master/png250px/{code}.png'
-                country_dir = Path(__file__).resolve().parent / "_codes" / "country.csv"
-                codes = pd.read_csv(country_dir, index_col='country')['code'].to_dict()
-                image_dict = {}
-                for country in self.df_values.columns:
-                    code = codes[country.lower()]
-                    image_dict[country] = mimage.imread(url.format(code=code))
-                return image_dict
+            from ._utils import read_images
+            return read_images(images.lower(), self.df_values.columns)
         if len(images) != len(self.df_values.columns):
             raise ValueError('The number of images does not match the number of columns')
         if isinstance(images, list):
@@ -323,6 +327,7 @@ class _LineChartRace(CommonChart):
             return
         ax = self.fig.axes[0]
         s = self.df_values.iloc[i]
+        s_all = self.all_values.iloc[i]
         x, y = s.name, s.to_dict()
         if self.is_x_date:
             x = mdates.date2num(x)
@@ -339,8 +344,8 @@ class _LineChartRace(CommonChart):
             y['___agg_line___'] = self.agg_line.iloc[i]
             visible['___agg_line___'] = True
 
-        if isinstance(self.df_others, pd.Series):
-            y['___others_line___'] = self.df_others.iloc[i]
+        if self.others_agg_line is not None:
+            y['___others_line___'] = self.others_agg_line.iloc[i]
             visible['___others_line___'] = True
 
         for col, collection in self.collections.items():
@@ -382,7 +387,7 @@ class _LineChartRace(CommonChart):
                 collection.set_segments(seg)
 
         if self.period_summary_func:
-            text_dict = self.add_period_summary(ax, s)
+            text_dict = self.add_period_summary(ax, s_all)
             text = self.texts['__period_summary_func__']
             x_period, y_period, text_val = text_dict.pop('x'), text_dict.pop('y'), text_dict.pop('s')
             text.set_position((x_period, y_period))
@@ -401,12 +406,12 @@ class _LineChartRace(CommonChart):
                 vis = visible[col]
                 img.set_visible(vis)
 
-        if self.tick_template:
-            ax.yaxis.set_major_formatter(self.tick_template)
-
     def init_func(self):
         ax = self.fig.axes[0]
         s = self.df_values.iloc[0] # current Series
+        s_all = self.all_values.iloc[0]
+        if len(self.df_others) > 0:
+            s_others = self.df_others.iloc[0]
         x, y = s.name, s.to_dict()
         if self.is_x_date:
             x = mdates.date2num(x)
@@ -423,11 +428,15 @@ class _LineChartRace(CommonChart):
             val = y[col]
             vis = visible[col]
             color = self.colors[col]
-            text = ax.text(x + x_extra, val, col, ha='left', va='center', size='smaller', visible=vis)
-            lw = 1.5
+            text = ax.text(x + x_extra, val, col, visible=vis, **self.line_label_font)
+            lw = self.line_kwargs['lw']
+            ls = self.line_kwargs['ls']
+            alpha = self.line_kwargs.get('alpha', 1)
+            color[-1] = alpha
             if self.line_width_data is not None:
                 lw = self.line_width_data.iloc[0][col]
-            collection = ax.add_collection(LineCollection([[(x, val)]], colors=[color], visible=vis, linewidths=[lw]))
+            lc = LineCollection([[(x, val)]], colors=[color], visible=vis, linewidths=[lw], linestyles=[ls])
+            collection = ax.add_collection(lc)
             self.texts[col] = text
             self.collections[col] = collection
 
@@ -441,23 +450,27 @@ class _LineChartRace(CommonChart):
             color = self.agg_line_kwargs['color']
             lw = self.agg_line_kwargs.get('lw')
             ls = self.agg_line_kwargs.get('ls')
+            alpha = self.agg_line_kwargs.get('alpha', AGG_COLOR[-1])
+            color = tuple(color[:3]) + (alpha,)
             label = self.agg_line_label
             val = self.agg_line.iloc[0]
             lc = LineCollection([[(x, val)]], colors=[color], linewidths=[lw], linestyles=[ls])
             collection = ax.add_collection(lc)
-            text = ax.text(x + x_extra, val, label, ha='left', va='center', size='smaller')
+            text = ax.text(x + x_extra, val, label, **self.line_label_font)
 
             label = '___agg_line___'
             self.collections[label] = collection
             self.texts[label] = text
             self.colors[label] = color
 
-        if isinstance(self.df_others, pd.Series):
+        if self.others_agg_line is not None:
             color = self.others_line_kwargs['color']
             lw = self.others_line_kwargs.get('lw')
             ls = self.others_line_kwargs.get('ls')
+            alpha = self.others_line_kwargs.get('alpha', OTHERS_COLOR[-1])
+            color = tuple(color[:3]) + (alpha,)
             label = self.others_line_label
-            val = self.df_others.iloc[0]
+            val = self.others_agg_line.iloc[0]
             lc = LineCollection([[(x, val)]], colors=[color], linewidths=[lw], linestyles=[ls])
             collection = ax.add_collection(lc)
             text = ax.text(x + x_extra, val, label, ha='left', va='center', size='smaller')
@@ -468,7 +481,7 @@ class _LineChartRace(CommonChart):
             self.colors[label] = color
 
         if self.period_summary_func:
-            text_dict = self.add_period_summary(ax, s)
+            text_dict = self.add_period_summary(ax, s_all)
             text = ax.text(transform=ax.transAxes, **text_dict)
             self.texts['__period_summary_func__'] = text
 
@@ -535,7 +548,7 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
                     images=None, colors=None, title=None, line_label_font=None, 
                     tick_label_font=None, tick_template='{x:,.0f}', shared_fontdict=None, 
                     scale='linear', fig=None, writer=None, line_kwargs=None, 
-                    fig_kwargs=None, filter_column_colors=False):
+                    fig_kwargs=None):
     '''
     Create an animated line chart race using matplotlib. Data must be in 
     'wide' format where each row represents a single time period and each 
@@ -545,9 +558,9 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
     If no `filename` is given, an HTML string is returned, otherwise the 
     animation is saved to disk.
 
-    You must have ffmpeg installed on your machine to save files to disk.
-    Get ffmpeg here: https://www.ffmpeg.org/download.html. To save .gif 
-    files, install ImageMagick.
+    You must have ffmpeg installed on your machine to save videos to disk
+    and ImageMagick to save animated gifs. Read more here:
+    https://www.dexplo.org/bar_chart_race/installation/
 
     Parameters
     ----------
@@ -558,10 +571,9 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         The index can be of any type.
 
     filename : `None` or str, default None
-        If `None` return animation as an HTML5 string.
-        If a string, save animation to that filename location. 
-        Use .mp4, .gif, .html, .mpeg, .mov and any other extensions supported
-        by ffmpeg or ImageMagick.
+        If `None` return animation as an HTML5 string. If a string, save 
+        animation to that filename location. Use .mp4, .gif, .html, .mpeg, 
+        .mov or any other extensions supported by ffmpeg or ImageMagick.
 
     n_lines : int, default None
         The maximum number of lines to display on the graph. 
@@ -570,15 +582,15 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         period. 
         
         See the others_line_func parameter for more options on plotting 
-        the other columns outside of the top n_lines.
-        By default, use all lines.
+        the other columns outside of the top n_lines. Use all lines by default.
 
     steps_per_period : int, default 10
         The number of steps to go from one time period to the next. 
+        Increasing this number creates smoother animations.
 
     period_length : int, default 500
         Number of milliseconds to animate each period (row). 
-        Default is 500ms (half of a second)
+        Default is 500ms (half of a second).
 
     end_period_pause : int, default 0
         Number of milliseconds to pause the animation at the end of
@@ -591,20 +603,19 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         to be a pause. The pause will be in increments of this
         calculated interval and not exact. For example, setting the
         end_period_pause to 725 will produce a pause of 700 
-        milliseconds (by default).
+        milliseconds when using the defaults.
 
     period_summary_func : function, default None
         Custom text added to the axes each period.
         Create a user-defined function that accepts one pandas Series of the 
-        current time period's values. It must return a dictionary 
-        containing at a minimum the keys "x", "y", and "s" which will be 
-        passed to the matplotlib `text` method.
-
+        current time period's values. It must return a dictionary containing 
+        the keys "x", "y", and "s" which will be passed to the matplotlib 
+        `text` method.
         Example:
         def func(values):
             total = values.sum()
             s = f'Worldwide deaths: {total}'
-            return {'x': .85, 'y': .2, 's': s, 'ha': 'right', 'size': 11}
+            return {'x': .05, 'y': .85, 's': s, 'size': 10}
 
     line_width_data : DataFrame, default None
         Control the width of the line at each time period.
@@ -617,15 +628,16 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         a string that the DataFrame `agg` method understands or a user-defined 
         function. Use agg_line_kwargs to style the line and provide it a label.
 
-        If providing function, it will be passed all values of the current period 
-        as a Series. Return a single value that summarizes the current period.
+        If providing function, it will be passed all values of the current 
+        period as a Series. Return a single value that summarizes the current 
+        period.
 
         DataFrame agg strings - 'mean', 'median', 'max', 'min', etc..
 
     agg_line_kwargs : dict, default None
         A dictionary of matplotlib line properties used with agg_line_func. 
-        Use the key `s` to control the label of the line.
-        Keys `x` and `y` will be ignored as the position is already determined.
+        Use the key `s` to control the label of the line. Keys `x` and `y` 
+        will be ignored as the position is already determined.
         Example: 
         {
             's': 'Median',
@@ -639,18 +651,18 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         By default (None), these other lines will not be plotted. Use True to 
         plot the lines in a soft gray color without labels or images.
 
-        Aggregate all of the other column values by passing in a string that the 
-        DataFrame agg method understands or a user-defined function, which will
-        be passed a pandas Series of the current period.
-        
+        Aggregate all of the other column values by passing in a string that 
+        the DataFrame agg method understands or a user-defined function, which 
+        will be passed a pandas Series of just these other values for this 
+        time period.
         Example:
         def my_others_line_func(s):
             return s.median()
 
     others_line_kwargs : dict, default None
         A dictionary of matplotlib line properties used with others_line_func. 
-        Use the key `s` to control the label of the line.
-        Keys `x` and `y` will be ignored as the position is already determined.
+        Use the key `s` to control the label of the line. Keys `x` and `y` 
+        will be ignored as the position is already determined.
         Example: 
         {
             's': 'Rest of World',
@@ -660,22 +672,27 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         }
 
     fade : float, default 1
-        Used to slowly fade historical values of the line, i.e. decrease the 
+        Use to slowly fade historical values of the line, i.e. decrease the 
         opacity (alpha). This number multiplies the current alpha of the line 
         each time period.
         
         Choose a number between 0 and 1. When 1, no fading occurs. 
-        This number will likely need to be very close to 1, as alpha will
-        quickly go to zero
+        This number will likely need to be close to 1, as alpha decreases fast.
         
     min_fade : float, default .3
-        Minimum value of alpha for each line. Choose a number between 0 and 1.
-        Use 0 to have the line eventually become completely transparent.
+        Minimum value of alpha for each line when setting fade < 1. Choose a 
+        number between 0 and 1. Use 0 to have the line eventually become 
+        completely transparent.
     
-    images : list or dict, default None
-        Images to use for the end of the line. Provide a list of filenames
-        or URLs where the images are located in the same order as the columns of
-        the DataFrame or a dictionary with keys as the column names.
+    images : str, list, or dict, default None
+        Images to use for the end of the line. Use a string to use one of the
+        built-in image databases by name. Currently, there are two image bases
+        'country', and 'nfl'. If your columns are countries are NFL teams, the 
+        images will automatically be found.
+        
+        Otherwise, provide a list of filenames or URLs where the images are 
+        located in the same order as the columns of the DataFrame or a 
+        dictionary mapping column names to filenames/URLs.
 
     colors : str, matplotlib colormap instance, or list of colors, default 'dark12'
         Colors to be used for the lines. All matplotlib and plotly 
@@ -690,7 +707,7 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
 
     title : str or dict, default None
         Title of plot as a string. Use a dictionary to supply several title 
-        parameters. You must use the key 'label' for the title.
+        parameters. You must use the key 'label' for the text.
         Example:
         {
             'label': 'My Line Chart Race Title',
@@ -709,21 +726,19 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         Use a dictionary to supply several font properties.
 
     tick_template : str or function, default '{x:,.0f}'
-        Formats the ticks on the axis with numeric values 
-        (x-axis when horizontal and y-axis when vertical). If given a string,
-        pass it to the ticker.StrMethodFormatter matplotlib function. 
+        Formats the ticks on the y-axis with numeric values When given a 
+        string, it's passed to the ticker.StrMethodFormatter function. 
         Use 'x' as the variable
         Example: '{x:10.2f}'
 
-        If given a function, its passed to ticker.FuncFormatter, which
-        implicitly passes it two variables `x` and `pos` and must return
+        WHen given a function, it's passed to ticker.FuncFormatter, which
+        implicitly passes it two parameters `x` and `pos` and must return
         a string.
 
     shared_fontdict : dict, default None
         Dictionary of font properties shared across the tick labels, 
         line labels, and title. The only property not shared is `size`. 
         It will be ignored if you try to set it.
-
         Possible keys are:
             'family', 'weight', 'color', 'style', 'stretch', 'weight', 'variant'
         Example:
@@ -734,10 +749,11 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         }
 
     scale : 'linear' or 'log', default 'linear'
-        Type of scaling to use for the axis containing the values
+        Type of scaling to use for the y-axis
 
     fig : matplotlib Figure, default None
-        For greater control over the aesthetics, supply your own figure.
+        For greater control over the aesthetics, supply your own figure
+        with at least one axes.
 
     writer : str or matplotlib Writer instance
         This argument is passed to the matplotlib FuncAnimation.save method.
@@ -750,17 +766,13 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         >>> from matplotlib import animation
         >>> animation.writers.list()
 
-        You must have ffmpeg or ImageMagick installed in order
-
-    line_kwargs : dict, default `None` (alpha=.8)
-        Other keyword arguments (within a dictionary) forwarded to the 
-        matplotlib Line2D function. If no value for 'alpha' is given,
-        then it is set to .8 by default.
+    line_kwargs : dict, default None
+        Other keyword arguments within a dictionary used to control line
+        properties. 
         Sample properties:
             `lw` - Line width, default is 1.5
             'ls' - Line style, '-', '--', '-.', ':'
             `alpha` - opacity of line, 0 to 1
-            'ms' - marker style
 
     fig_kwargs : dict, default None
         A dictionary of keyword arguments passed to the matplotlib
@@ -772,31 +784,6 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
             'dpi': 120,
             'facecolor': 'red'
         }
-
-    filter_column_colors : bool, default `False`
-        When setting n_lines, it's possible that some columns never 
-        appear in the animation. Regardless, all columns get assigned
-        a color by default. 
-        
-        For instance, suppose you have 100 columns 
-        in your DataFrame, set n_lines to 10, and 15 different columns 
-        make at least one appearance in the animation. Even if your 
-        colormap has at least 15 colors, it's possible that many 
-        lines will be the same color, since each of the 100 columns is
-        assigned of the colormaps colors.
-
-        Setting this to `True` will map your colormap to just those 
-        columns that make an appearance in the animation, helping
-        avoid duplication of colors.
-
-        Setting this to `True` will also have the (possibly unintended)
-        consequence of changing the colors of each color every time a 
-        new integer for n_lines is used.
-
-        EXPERIMENTAL
-        This parameter is experimental and may be changed/removed
-        in a later version.
-
 
     Returns
     -------
@@ -836,8 +823,7 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
         fig=None, 
         writer=None, 
         line_kwargs={'alpha': .7},
-        fig_kwargs={'figsizse': (6, 3.5), 'dpi': 144},
-        filter_column_colors=False)        
+        fig_kwargs={'figsizse': (6, 3.5), 'dpi': 144})        
 
     Font Help
     ---------
@@ -850,5 +836,5 @@ def line_chart_race(df, filename=None, n_lines=None, steps_per_period=10,
                          period_summary_func, line_width_data, agg_line_func, agg_line_kwargs, 
                          others_line_func, others_line_kwargs, fade, min_fade, images, colors, 
                          title, line_label_font, tick_label_font, tick_template, shared_fontdict, 
-                         scale, fig, writer, line_kwargs, fig_kwargs, filter_column_colors)
+                         scale, fig, writer, line_kwargs, fig_kwargs)
     return lcr.make_animation()