+++ bar_chart_race/_pandas_accessor.py
@@ -3,7 +3,7 @@ import re
 import pandas as pd
 
 from ._bar_chart_race import bar_chart_race as bcr
-from ._bar_chart_race_plotly import bar_chart_race_plotly as bcrp
+from ._line_chart_race import line_chart_race as lcr
 from ._utils import prepare_wide_data as pwd, prepare_long_data as pld
 
 
@@ -14,37 +14,35 @@ class _BCR:
         self._df = df
 
     def bar_chart_race(self, filename=None, orientation='h', sort='desc', n_bars=None, 
-                       fixed_order=False, fixed_max=False, steps_per_period=10, 
-                       period_length=500, end_period_pause=0, interpolate_period=False, 
-                       bar_label_position='outside', bar_label_fmt='{x:,.0f}',
-                       bar_size=.95, period_label=True, period_fmt=None, 
-                       period_summary_func=None, perpendicular_bar_func=None, 
-                       figsize=(6, 3.5), cmap=None, title=None, bar_label_size=7, 
-                       tick_label_size=7, shared_fontdict=None, scale='linear', writer=None, 
-                       fig=None, dpi=144, bar_kwargs=None, filter_column_colors=False):
+                   fixed_order=False, fixed_max=False, steps_per_period=10, 
+                   period_length=500, end_period_pause=0, interpolate_period=False, 
+                   period_label=True, period_template=None, period_summary_func=None,
+                   perpendicular_bar_func=None, colors=None, title=None, bar_size=.95,
+                   bar_textposition='outside', bar_texttemplate='{x:,.0f}',
+                   bar_label_font=None, tick_label_font=None, tick_template='{x:,.0f}',
+                   shared_fontdict=None, scale='linear', fig=None, writer=None, 
+                   bar_kwargs=None,  fig_kwargs=None, filter_column_colors=False):
 
         return bcr(self._df, filename, orientation, sort, n_bars, fixed_order, fixed_max,
-                  steps_per_period, period_length, end_period_pause, interpolate_period, 
-                   bar_label_position, bar_label_fmt, bar_size, period_label, period_fmt, 
-                   period_summary_func, perpendicular_bar_func, figsize, cmap, title, 
-                   bar_label_size, tick_label_size, shared_fontdict, scale, writer, fig, 
-                   dpi, bar_kwargs, filter_column_colors)
+                        steps_per_period, period_length, end_period_pause, interpolate_period, 
+                        period_label, period_template, period_summary_func, perpendicular_bar_func,
+                        colors, title, bar_size, bar_textposition, bar_texttemplate, 
+                        bar_label_font, tick_label_font, tick_template, shared_fontdict, scale, 
+                        fig, writer, bar_kwargs, fig_kwargs, filter_column_colors)
 
-    def bar_chart_race_plotly(self, filename=None, orientation='h', sort='desc', n_bars=None, 
-                              fixed_order=False, fixed_max=False, steps_per_period=10, 
-                              period_length=500, interpolate_period=False, period_label=True, 
-                              period_fmt=None, period_summary_func=None, perpendicular_bar_func=None,
-                              colors=None, title=None, bar_size=.95, textposition='outside', 
-                              texttemplate=None, bar_label_font=12, tick_label_font=12, 
-                              hovertemplate=None, slider=True, scale='linear', bar_kwargs=None, 
-                              layout_kwargs=None, write_html_kwargs=None, filter_column_colors=False):
-
-        return bcrp(self._df, filename, orientation, sort, n_bars, fixed_order, fixed_max,
-                    steps_per_period, period_length, interpolate_period, period_label, 
-                    period_fmt, period_summary_func, perpendicular_bar_func, colors, title, 
-                    bar_size, textposition, texttemplate, bar_label_font, tick_label_font, 
-                    hovertemplate, slider, scale, bar_kwargs, layout_kwargs, write_html_kwargs,
-                    filter_column_colors)
+    def line_chart_race(self, filename=None, n_lines=None, steps_per_period=10, 
+                        period_length=500, end_period_pause=0, period_summary_func=None, 
+                        line_width_data=None, agg_line_func=None, agg_line_kwargs=None, 
+                        others_line_func=None, others_line_kwargs=None, fade=1, min_fade=.3, 
+                        images=None, colors=None, title=None, line_label_font=None, 
+                        tick_label_font=None, tick_template='{x:,.0f}', shared_fontdict=None, 
+                        scale='linear', fig=None, writer=None, line_kwargs=None, 
+                        fig_kwargs=None):
+        return lcr(self._df, filename, n_lines, steps_per_period, period_length, end_period_pause, 
+                    period_summary_func, line_width_data, agg_line_func, agg_line_kwargs, 
+                    others_line_func, others_line_kwargs, fade, min_fade, images, colors, 
+                    title, line_label_font, tick_label_font, tick_template, shared_fontdict, 
+                    scale, fig, writer, line_kwargs, fig_kwargs)
 
     def prepare_wide_data(self, orientation='h', sort='desc', n_bars=None, interpolate_period=False, 
                           steps_per_period=10, compute_ranks=True):
@@ -59,6 +57,29 @@ class _BCR:
 
 
 _BCR.bar_chart_race.__doc__ = re.sub('df : .*(?=filename :)', '',  bcr.__doc__, flags=re.S)
-_BCR.bar_chart_race_plotly.__doc__ = re.sub('df : .*(?=filename :)', '',  bcrp.__doc__, flags=re.S)
+_BCR.line_chart_race.__doc__ = re.sub('df : .*(?=filename :)', '',  lcr.__doc__, flags=re.S)
 _BCR.prepare_wide_data.__doc__ = re.sub('df : .*(?=filename :)', '',  pwd.__doc__, flags=re.S)
 _BCR.prepare_long_data.__doc__ = re.sub('df : .*(?=filename :)', '',  pld.__doc__, flags=re.S)
+
+import importlib
+if importlib.util.find_spec('plotly'):
+    from ._bar_chart_race_plotly import bar_chart_race_plotly as bcrp
+    def bar_chart_race_plotly(self, filename=None, orientation='h', sort='desc', n_bars=None, 
+                          fixed_order=False, fixed_max=False, steps_per_period=10, 
+                          period_length=500, end_period_pause=0, interpolate_period=False, 
+                          period_label=True, period_template=None, period_summary_func=None, 
+                          perpendicular_bar_func=None, colors=None, title=None, bar_size=.95, 
+                          bar_textposition='outside', bar_texttemplate=None, bar_label_font=None, 
+                          tick_label_font=None, hovertemplate=None, slider=True, scale='linear', 
+                          bar_kwargs=None, layout_kwargs=None, write_html_kwargs=None, 
+                          filter_column_colors=False):
+
+        return bcrp(self._df, filename, orientation, sort, n_bars, fixed_order, fixed_max,
+                        steps_per_period, period_length, end_period_pause, interpolate_period, 
+                        period_label, period_template, period_summary_func, perpendicular_bar_func, 
+                        colors, title, bar_size, bar_textposition, bar_texttemplate, bar_label_font, 
+                        tick_label_font, hovertemplate, slider, scale, bar_kwargs, layout_kwargs, 
+                        write_html_kwargs, filter_column_colors)
+
+    setattr(_BCR, 'bar_chart_race_plotly', bar_chart_race_plotly)
+    _BCR.bar_chart_race_plotly.__doc__ = re.sub('df : .*(?=filename :)', '',  bcrp.__doc__, flags=re.S)