+++ docs/whats_new.md
@@ -2,9 +2,23 @@
 
 ## Version 0.2
 
-* Bar label position able to be specified ('outside', 'inside', or None) using new parameter `bar_label_position`.
-* Bar label formatting possible with string or function using new parameter `bar_label_fmt`
-* Removal of parameter `label_bars`
+Upcoming release on July xx, 2020
+
+### Major New Features
+
+* Plotly animated bar charts with `bar_chart_race_plotly`
+* Line chart races with `line_chart_race`
+
+#### Other enhancements
+
+* Integration directly into pandas DataFrames - `df.bcr.bar_chart_race`
+* Bar label position able to be specified ('outside', 'inside', or None) using new parameter `bar_textposition`
+* Bar label formatting possible with string or function using new parameter `bar_texttemplate`
+* Added `end_period_pause` parameter that creates a pause (in milliseconds) at the end of each period
+* Parameter `title`, in addition to a string, can also be a dictionary using `'label'` as the key for the title. Other keys may be used to control text properties
+* Removed parameters `figsize` and `dpi` in favor of `fig_kwargs` dictionary capable of taking all matplotlib `Figure` parameters
+* Figure background color able to be saved
+* Several parameters changed name and order
 
 ## Version 0.1